% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey.R
\name{json2sf_tapi}
\alias{json2sf_tapi}
\title{Convert output from transportapi.com into sf object}
\usage{
json2sf_tapi(obj, apitype)
}
\arguments{
\item{obj}{Object from transportapi.com read-in with}

\item{apitype}{Type of routing can be car, cycle, public (DEFAULT)}
}
\description{
Convert output from transportapi.com into sf object
}
\examples{
None

}
