% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_station.R
\name{train_station}
\alias{train_station}
\title{Get train station info from transportAPI}
\usage{
train_station(station_code, live = FALSE, date = NULL, time = NULL,
  called_at = NULL, calling_at = NULL, destination = NULL,
  from_offset = NULL, to_offset = NULL, operator = NULL,
  origin = NULL, service = NULL, station_detail = NULL,
  train_status = NULL, type = NULL, silent = TRUE,
  app_id = Sys.getenv("TRANSPORTAPI_app_id"),
  app_key = Sys.getenv("TRANSPORTAPI_app_key"),
  base_url = "http://transportapi.com/", save_raw = FALSE)
}
\arguments{
\item{station_code}{The station code of the station of interest. Can be CRS or TIPLOC.}

\item{live}{Logical, use the live API}

\item{date}{The date of interest in yyyy-mm-dd format.}

\item{time}{The time of interest in hh:mm format.}

\item{called_at}{Only include services that call at the given station, before calling at the station of interest.}

\item{calling_at}{Only include services that call at the given station, after calling at the station of interest.}

\item{destination}{Only include services terminating at the given station.}

\item{from_offset}{Modifies the start of the time window for which services are retrieved. By default, this is coincides with the date/time of interest.}

\item{to_offset}{Modifies the end of the time window for which services are retrieved. By default, this is two hours in the future relative to the date/time of interest.}

\item{operator}{Only include services that are operated by the given operator.}

\item{origin}{Only include services originating from the given station.}

\item{service}{Only include services that have the given service code.}

\item{station_detail}{Add additional detail for the services returned at certain stations.}

\item{train_status}{Only include services having the specified train status.}

\item{silent}{Logical (default is FALSE). TRUE hides request sent.}

\item{app_id}{The app id used. By default this uses \code{Sys.getenv("TRANSPORTAPI_app_id")}.}

\item{app_key}{The app key used. By default this uses \code{Sys.getenv("TRANSPORTAPI_app_key")}.}

\item{base_url}{The base url from which to construct API requests
(with default set to main server)}

\item{save_raw}{Boolean value which returns raw list from the json if TRUE (FALSE by default).}
}
\description{
R interface to transportAPI
See \href{https://developer.transportapi.com/docs}{developer.transportapi.com/docs} for details.
}
\details{
Requires the internet and a transportapi.com API key.
transportapi.com does not yet work worldwide.

You need to have an api key and app id for this code to run.
By default it uses the TRANSPORTAPI_app_id and  TRANSPORTAPI_app_key environment variable.
This can be set with \code{usethis::edit_r_environ()}.
}
