% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey.R
\name{journey.batch}
\alias{journey.batch}
\title{Wrapper for jounrey fuction to perform multiple routings}
\usage{
journey.batch(from, to, ...)
}
\arguments{
\item{from}{Longitude/Latitude pair, e.g. \code{c(-0.134649,51.529258)} or SF points of class "sfc_POINT" "sfc" and length one}

\item{to}{Longitude/Latitude pair, e.g. \code{c(-0.088780,51.506383)} or SF points of class "sfc_POINT" "sfc" and length one}
}
\description{
Wrapper for jounrey fuction to perform multiple routings
}
\examples{
None

}
